+++ charts/osm/templates/osm-bootstrap-deployment.yaml
@@ -1,23 +1,22 @@
-{{- if .Values.OpenServiceMesh.featureFlags.enableCRDConverter }}
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: osm-crd-converter
+  name: osm-bootstrap
   namespace: {{ include "osm.namespace" . }}
   labels:
     {{- include "osm.labels" . | nindent 4 }}
-    app: osm-crd-converter
+    app: osm-bootstrap
     meshName: {{ .Values.OpenServiceMesh.meshName }}
 spec:
   replicas: {{ .Values.OpenServiceMesh.crdConverter.replicaCount }}
   selector:
     matchLabels:
-      app: osm-crd-converter
+      app: osm-bootstrap
   template:
     metadata:
       labels:
         {{- include "osm.labels" . | nindent 8 }}
-        app: osm-crd-converter
+        app: osm-bootstrap
   {{- if .Values.OpenServiceMesh.crdConverter.podLabels }}
   {{- toYaml .Values.OpenServiceMesh.crdConverter.podLabels | nindent 8 }}
   {{- end }}
@@ -33,8 +32,8 @@ spec:
         kubernetes.io/arch: amd64
         kubernetes.io/os: linux
       containers:
-        - name: osm-crd-converter
-          image: "{{ .Values.OpenServiceMesh.image.registry }}/osm-crd-converter:{{ .Values.OpenServiceMesh.image.tag }}"
+        - name: osm-bootstrap
+          image: "{{ .Values.OpenServiceMesh.image.registry }}/osm-bootstrap:{{ .Values.OpenServiceMesh.image.tag }}"
           imagePullPolicy: {{ .Values.OpenServiceMesh.image.pullPolicy }}
           ports:
             - name: "tls"
@@ -43,7 +42,7 @@ spec:
               containerPort: 9091
             - name: "health"
               containerPort: 9095
-          command: ['/osm-crd-converter']
+          command: ['/osm-bootstrap']
           args: [
             "--verbosity", "{{.Values.OpenServiceMesh.controllerLogLevel}}",
             "--osm-namespace", "{{ include "osm.namespace" . }}",
@@ -80,8 +79,8 @@ spec:
               path: /healthz
               port: 9095
           env:
-            # The CRD_CONVERTER_POD_NAME env variable sets pod name dynamically, used by osm-crd-converter to register events
-            - name: CRD_CONVERTER_POD_NAME
+            # The BOOTSTRAP_POD_NAME env variable sets pod name dynamically, used by osm-bootstrap to register events
+            - name: BOOTSTRAP_POD_NAME
               valueFrom:
                 fieldRef:
                   fieldPath: metadata.name
@@ -93,4 +92,3 @@ spec:
       tolerations:
 {{ toYaml .Values.OpenServiceMesh.controlPlaneTolerations | indent 8 }}
     {{- end }}
-{{- end }}
