+++ cmd/osm-bootstrap/osm-bootstrap_test.go
@@ -103,6 +103,17 @@ func TestValidateCLIParams(t *testing.T) {
 				osmNamespace = "valid-ns"
 			},
 			verify: func(err error) {
+				assert.NotNil(err)
+				assert.Contains(err.Error(), "--ca-bundle-secret-name")
+			},
+		},
+		{
+			caseName: "osm-namespace and ca-bundle-secret-name is valid",
+			setup: func() {
+				osmNamespace = "valid-ns"
+				caBundleSecretName = "valid-ca-bundle"
+			},
+			verify: func(err error) {
 				assert.Nil(err)
 			},
 		},
