--- dockerfiles/Dockerfile.osm-crd-converter
@@ -1,2 +1,0 @@
-FROM gcr.io/distroless/static
-COPY osm-crd-converter /
